+++ src/globalTypes.d.ts
@@ -1,0 +1,10 @@
+//@ts-ignore
+export type Actor5e = globalThis.dnd5e.documents.Actor5e
+//@ts-ignore
+export type TokenDocument5e = globalThis.dnd5e.documents.TokenDocument5e
+//@ts-ignore
+export type Item5e = globalThis.dnd5e.documents.Item5e
+//@ts-ignore
+export type ActiveEffect5e = globalThis.dnd5e.documents.ActiveEffect5e
+
+export type ChatMessage = { content: string, update(any) }